[refactor] Moved norllm into llm folder (#74)
