+++ backend/src/services/llm/norLLM/request.ts
@@ -1,4 +1,4 @@
-import { parsePdf } from "../middleware/PdfParser.js";
+import { parsePdf } from "../../../middleware/PdfParser.js";
 
 const url = "https://llm.hpc.ntnu.no/v1/chat/completions";
